Improve UI of Crawl, Main page with header
